requirements.txt
Flask>=0.10.1
Jinja2>=2.7.3
MarkupSafe==0.23
PyRSS2Gen==1.0.0
Twisted==12.2.0
Werkzeug==0.9.6
altgraph==0.10.1
bdist-mpkg==0.4.4
beautifulsoup4==4.3.2
bonjour-py==0.3
certifi==14.05.14
itsdangerous==0.24
macholib==1.5
matplotlib==1.1.1
modulegraph==0.10.1
numpy==1.6.2
py2app==0.7.1
pyOpenSSL==0.13
python-dateutil==1.5
python-googlegeocoder==0.2.0
pytz==2012d
requests==2.3.0
scipy==0.11.0
six==1.4.1
virtualenv==1.11.6
wsgiref==0.1.2
xattr==0.6.4
zope.interface==3.8.0